feat: método criar chamado funcionando, separando método de criar chamado na classe ChamadoService e conversor de enum para string por conta de conflito com o banco :sparkles:
