+++ sime.api/src/main/java/br/com/sime/api/controllers/UsuarioController.java
@@ -1,10 +1,7 @@
 package br.com.sime.api.controllers;
 
-import br.com.sime.api.DTOs.ChamadoRequestDTO;
 import br.com.sime.api.DTOs.LoginDTO;
-import br.com.sime.api.entities.chamados.Chamado;
 import br.com.sime.api.services.UsuarioService;
-import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -29,11 +26,5 @@ public class UsuarioController {
     public ResponseEntity<Boolean> login(@RequestBody LoginDTO dto) {
         usuarioService.login(dto.getRmUsuario(), dto.getSenhaUsuario(), dto.getIdTipoPerfil(), dto.getCodEscola());
         return new ResponseEntity<>(true, HttpStatus.OK);
-    }
-
-    @PostMapping("/{rmUsuario}/chamado")
-    public ResponseEntity<?> criarChamado(@PathVariable String rmUsuario, @Valid @RequestBody ChamadoRequestDTO ChamadoDTO) {
-        Chamado newChamado = usuarioService.criarChamado(rmUsuario, ChamadoDTO);
-        return new ResponseEntity<>(newChamado, HttpStatus.CREATED);
     }
 }
