+++ sime.api/src/main/java/br/com/sime/api/services/UsuarioService.java
@@ -1,12 +1,7 @@
 package br.com.sime.api.services;
 
-import br.com.sime.api.DTOs.ChamadoRequestDTO;
-import br.com.sime.api.entities.chamados.Chamado;
-import br.com.sime.api.entities.chamados.TipoChamado;
 import br.com.sime.api.entities.outros.TipoPerfil;
 import br.com.sime.api.entities.usuarios.Usuario;
-import br.com.sime.api.enums.PrioridadeChamadoEnum;
-import br.com.sime.api.enums.StatusChamadoEnum;
 import br.com.sime.api.exceptions.NotFoundException;
 import br.com.sime.api.exceptions.SenhaIncorretaException;
 import br.com.sime.api.repositories.ChamadoRepository;
@@ -15,7 +10,6 @@ import br.com.sime.api.repositories.TipoPerfilRepository;
 import br.com.sime.api.repositories.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 
 @Service
@@ -59,26 +53,6 @@ public class UsuarioService {
                     return true;
                 })
                 .orElseThrow(() -> new NotFoundException("Usuário não encontrado", "Usuário com RM: " + rmUsuario + " não encontrado"));
-    }
-
-    public Chamado criarChamado(String rmUsuario, ChamadoRequestDTO dto) {
-        Usuario usuario = usuarioRepository.findByRmUsuario(rmUsuario)
-                .orElseThrow(() -> new NotFoundException("Usuário não encontrado", "Usuário não encontrado: " + rmUsuario));
-
-        TipoChamado tipoChamado = tipoChamadoRepository.findByNomeTipoChamadoIgnoreCase(dto.getTipoChamado())
-                .orElseThrow(() -> new NotFoundException("Tipo de chamado não encontrado", "Tipo: " + dto.getTipoChamado()));
-
-        Chamado chamado = new Chamado();
-        chamado.setTituloChamado(dto.getTituloChamado());
-        chamado.setDescChamado(dto.getDescChamado());
-        chamado.setLocalChamado(dto.getLocalChamado());
-        chamado.setUsuario(usuario);
-        chamado.setStatusChamado(StatusChamadoEnum.PENDENTE);
-        chamado.setImgChamado(dto.getImgChamado());
-        chamado.setTipoChamado(tipoChamado);
-        //chamado.setPrioridadeChamado(PrioridadeChamadoEnum.ALTA_PRIORIDADE); //Revisar
-
-        return chamadoRepository.save(chamado);
     }
 
     private TipoPerfil getTipoPerfilOrThrow(Long idTipoPerfil) {
