+++ sime.api/src/main/resources/application.properties
@@ -13,5 +13,6 @@ spring.jpa.show-sql=true
 spring.jpa.format-sql=true
 spring.jpa.properties.hibernate.id.new_generator_mappings=false
 spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+spring.jackson.serialization.write-dates-as-timestamps=false
 
 server.port=8080
